refactor: move entity commands to FSD APIs
