+++ src/entities/card/index.ts
@@ -1,3 +1,4 @@
+export { CardBulkMutationError, cardCommands } from "./api/commands";
 export { createCard } from "./model/card";
 export type { Card, CardEdit, CardId, CardNew, CardRaw, CardTextKey } from "./model/card";
 export { useCards } from "./hooks/useCards";
